+++ examples/python_rs/llm/vllm_disaggregated_prefill/monolith_worker.py
@@ -15,36 +15,35 @@
 
 
 import asyncio
+import uuid
+from parser import parse_vllm_args
 
 import uvloop
 import vllm
-from protocol import PrefillRequest, PrefillResponse
+from protocol import Request, Response
 from triton_distributed_rs import DistributedRuntime, triton_endpoint, triton_worker
 from vllm.engine.arg_utils import AsyncEngineArgs
 from vllm.logger import logger as vllm_logger
-from vllm.utils import FlexibleArgumentParser
 
 
-class VllmPrefillEngine:
+class VllmEngine:
     """
     Request handler for the generate endpoint
     """
 
     def __init__(self, engine_args: AsyncEngineArgs):
-        assert (
-            engine_args.kv_transfer_config.is_kv_producer
-        ), "Prefill worker must be a KV producer"
         self.engine = vllm.AsyncLLMEngine.from_engine_args(engine_args)
 
-    @triton_endpoint(PrefillRequest, PrefillResponse)
+    @triton_endpoint(Request, Response)
     async def generate(self, request):
-        vllm_logger.info(f"Received prefill request: {request}")
+        vllm_logger.debug(f"Received request: {request}")
         sampling_params = vllm.SamplingParams(**request.sampling_params)
+        request_id = str(uuid.uuid4())
         async for response in self.engine.generate(
-            request.prompt, sampling_params, request.request_id
+            request.prompt, sampling_params, request_id
         ):
             vllm_logger.debug(f"Generated response: {response}")
-            yield True
+            yield response.outputs[0].text
 
 
 @triton_worker()
@@ -53,16 +52,14 @@ async def worker(runtime: DistributedRuntime, engine_args: AsyncEngineArgs):
     Instantiate a `backend` component and serve the `generate` endpoint
     A `Component` can serve multiple endpoints
     """
-    component = runtime.namespace("triton-init").component("prefill")
+    component = runtime.namespace("triton-init").component("vllm")
     await component.create_service()
 
     endpoint = component.endpoint("generate")
-    await endpoint.serve_endpoint(VllmPrefillEngine(engine_args).generate)
+    await endpoint.serve_endpoint(VllmEngine(engine_args).generate)
 
 
 if __name__ == "__main__":
     uvloop.install()
-    parser = FlexibleArgumentParser()
-    parser = AsyncEngineArgs.add_cli_args(parser)
-    engine_args = AsyncEngineArgs.from_cli_args(parser.parse_args())
+    engine_args = parse_vllm_args()
     asyncio.run(worker(engine_args))
