+++ examples/python_rs/llm/vllm_disaggregated_prefill/prefill_worker.py
@@ -15,35 +15,37 @@
 
 
 import asyncio
-import uuid
 
 import uvloop
 import vllm
-from protocol import Request, Response
 from triton_distributed_rs import DistributedRuntime, triton_endpoint, triton_worker
 from vllm.engine.arg_utils import AsyncEngineArgs
 from vllm.logger import logger as vllm_logger
-from vllm.utils import FlexibleArgumentParser
+
+from .parser import parse_vllm_args
+from .protocol import PrefillRequest, PrefillResponse
 
 
-class VllmEngine:
+class VllmPrefillEngine:
     """
     Request handler for the generate endpoint
     """
 
     def __init__(self, engine_args: AsyncEngineArgs):
+        assert (
+            engine_args.kv_transfer_config.is_kv_producer
+        ), "Prefill worker must be a KV producer"
         self.engine = vllm.AsyncLLMEngine.from_engine_args(engine_args)
 
-    @triton_endpoint(Request, Response)
+    @triton_endpoint(PrefillRequest, PrefillResponse)
     async def generate(self, request):
-        vllm_logger.debug(f"Received request: {request}")
+        vllm_logger.info(f"Received prefill request: {request}")
         sampling_params = vllm.SamplingParams(**request.sampling_params)
-        request_id = str(uuid.uuid4())
         async for response in self.engine.generate(
-            request.prompt, sampling_params, request_id
+            request.prompt, sampling_params, request.request_id
         ):
             vllm_logger.debug(f"Generated response: {response}")
-            yield response.outputs[0].text
+            yield True
 
 
 @triton_worker()
@@ -52,16 +54,14 @@ async def worker(runtime: DistributedRuntime, engine_args: AsyncEngineArgs):
     Instantiate a `backend` component and serve the `generate` endpoint
     A `Component` can serve multiple endpoints
     """
-    component = runtime.namespace("triton-init").component("vllm")
+    component = runtime.namespace("triton-init").component("prefill")
     await component.create_service()
 
     endpoint = component.endpoint("generate")
-    await endpoint.serve_endpoint(VllmEngine(engine_args).generate)
+    await endpoint.serve_endpoint(VllmPrefillEngine(engine_args).generate)
 
 
 if __name__ == "__main__":
     uvloop.install()
-    parser = FlexibleArgumentParser()
-    parser = AsyncEngineArgs.add_cli_args(parser)
-    engine_args = AsyncEngineArgs.from_cli_args(parser.parse_args())
+    engine_args = parse_vllm_args()
     asyncio.run(worker(engine_args))
